# Optimize day 9