Local ∆ of 8.499 ms => 824.7 μs through a few changes

* Rather than starting from each element in the list in turn and scanning right, maintain a single sliding window.
* Use a more efficient "is this value a sum of values from the preamble" check. Note that the new implementation does discard some still-relevant pre-computed data step-to-step, so there may still be some juice to squeeze here.

# Before

```
time                 8.499 ms   (7.319 ms .. 9.851 ms)
                     0.919 R²   (0.881 R² .. 0.978 R²)
mean                 8.093 ms   (7.731 ms .. 8.588 ms)
std dev              1.154 ms   (832.1 μs .. 1.438 ms)
```

# After

```
time                 824.7 μs   (811.6 μs .. 841.9 μs)
                     0.996 R²   (0.993 R² .. 0.999 R²)
mean                 836.4 μs   (828.1 μs .. 850.1 μs)
std dev              36.72 μs   (24.88 μs .. 56.24 μs)
```